Move getLock() and removeLock() to a separate class
